Changed naming styles
* minor changes according to React best practices
