+++ src/components/info/Info.scss
@@ -1,0 +1,25 @@
+$primary-color: #64B5F6;
+
+.infoContainer {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    width: auto;
+    .paper {
+        display: flex;
+        flex-direction: column;
+        justify-content: center;
+        align-items: center;
+        height: auto;
+        width: 300px;
+        margin: 20px;
+        font-size: 20px;
+        color: $primary-color !important;
+        .buttonContainer {
+            display: flex;
+            flex-direction: row;
+            justify-content: flex-end;
+            width: 256px;
+        }
+    }
+}
