--- src/components/info/info.scss
@@ -1,27 +1,0 @@
-$primary-color: #64B5F6;
-
-.info-container {
-    display: flex;
-    flex-direction: column;
-    align-items: center;
-    width: auto;
-}
-
-.paper {
-    display: flex;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    height: auto;
-    width: 300px;
-    margin: 20px;
-    font-size: 20px;
-    color: $primary-color !important;
-}
-
-.button-container {
-    display: flex;
-    flex-direction: row;
-    justify-content: flex-end;
-    width: 256px;
-}
