+++ src/components/main-page/chart/ChartStyles.js
