--- src/components/main-page/main-page.scss
@@ -1,17 +1,0 @@
-.main-container {
-    display: flex;
-    flex-direction: column;
-    width: auto;
-    align-items: center;
-}
-
-.data-container {
-    display: flex;
-    width: 80%;
-    margin-top: 20px;
-    margin-bottom: 20px;
-}
-
-.chart-container {
-    height: 300px;
-}
