+++ src/components/main-page/main-table/MainTableStyles.js
