+++ src/components/main-page/timer/TimerStyles.js
